+++ frontend/src/web/app/components/AppLoader.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Loader } from '../../lib/fui/react';
-import logo from '../../common/app/images/logo.png';
+import { Loader } from '../../../lib/fui/react';
+import logo from '../../../common/app/images/logo.png';
 
 type Props = {
   visible: bool,