+++ frontend/src/web/app/components/index.js
@@ -1,3 +1,6 @@
+export { default as AppError } from './AppError';
+export { default as AppLoader } from './AppLoader';
+export { default as AppToaster } from './AppToaster';
 export { default as CenterPage } from './CenterPage';
-export { default as SinglePage } from './SinglePage';
 export { default as LocaleSelector } from './LocaleSelector';
+export { default as SinglePage } from './SinglePage';